all but not strtok

+++ s21_string.h
@@ -1,7 +1,7 @@
 #ifndef S21_STRING_H
 #define S21_STRING_H
 
-typedef unsigned int size_t;
-#define S21_NULL (void*)0
+typedef long unsigned int size_t;
+#define S21_NULL ((void*)0)
  
 #endif
